requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
flask>=3.0.0
gunicorn>=21.0.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tldextract>=5.0.0
pypdf>=4.0.0
python-docx>=1.0.0
lxml>=4.9.0
